refactor: Don't duplicate config types
