+++ src/renderer/src/components/Control/ControlAlteka.vue
@@ -60,7 +60,7 @@
 </template>
 
 <script setup lang="ts">
-import type { Alteka } from '@renderer/config'
+import type { Alteka } from '@shared/config'
 
 const alteka = defineModel<Alteka>({ required: true })
