+++ src/renderer/src/components/Control/ControlShare.vue
@@ -64,7 +64,7 @@
 
 <script setup lang="ts">
 import { computed, ref } from 'vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 
 const showShareDialog = defineModel<boolean>()
