+++ src/renderer/src/components/TestCard/AudioSyncCard.vue
@@ -129,7 +129,7 @@
 
 <script setup lang="ts">
 import { computed, ref, watch } from 'vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 const props = defineProps<{
