+++ src/renderer/src/components/TestCard/BarsHdrCard.vue
@@ -86,7 +86,7 @@
 
 <script setup lang="ts">
 import Swatch from './ColorSwatch.vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 defineProps<{
