+++ src/renderer/src/components/TestCard/BarsSimpleCard.vue
@@ -12,7 +12,7 @@
 
 <script setup lang="ts">
 import InfoCircle from './InfoCircle.vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 defineProps<{
