+++ src/renderer/src/components/TestCard/BarsSingleCard.vue
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import Swatch from './ColorSwatch.vue'
 import InfoCircle from './InfoCircle.vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 defineProps<{
