+++ src/renderer/src/components/TestCard/LedPanel.vue
@@ -20,7 +20,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 
 const props = defineProps<{
   config: Config
