+++ src/renderer/src/components/TestCard/LedWallCard.vue
@@ -24,7 +24,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import LedPanel from './LedPanel.vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 const props = defineProps<{
