+++ src/renderer/src/components/TestCard/PlaceholderCard.vue
@@ -50,7 +50,7 @@
 </template>
 
 <script setup lang="ts">
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 defineProps<{
