+++ src/renderer/src/components/TestCard/RampCard.vue
@@ -32,7 +32,7 @@
 import { computed } from 'vue'
 import Swatch from './ColorSwatch.vue'
 import InfoCircle from './InfoCircle.vue'
-import type { Config } from '@renderer/config'
+import type { Config } from '@shared/config'
 import type { Info } from '@renderer/views/Testcard.vue'
 
 const props = defineProps<{
