+++ tsconfig.web.json
@@ -3,6 +3,7 @@
   "include": [
     "src/renderer/src/env.d.ts",
     "src/renderer/src/**/*",
+    "src/shared/**/*",
     "src/renderer/src/**/*.vue",
     "src/preload/*.d.ts"
   ],
@@ -12,6 +13,9 @@
     "paths": {
       "@renderer/*": [
         "src/renderer/src/*"
+      ],
+      "@shared/*": [
+        "src/shared/*"
       ],
       "@assets/*": [
         "src/renderer/src/assets/*"
